+++ discogs/release_test.go
@@ -10,6 +10,6 @@ func TestGetRelease(t *testing.T) {
 	release, err := testClient.GetRelease(10201544)
 	assert.Nil(t, err)
 	assert.NotNil(t, release.ID)
-	assert.Equal(t, release.ID, 10201544)
+	assert.Equal(t, release.ID, int64(10201544))
 	assert.NotNil(t, release.Title)
 }
